Web_Version/requirements.txt
pillow~=11.0.0
pip~=24.3.1
wheel~=0.41.2
setuptools~=68.2.0
Django~=5.1.4
asgiref~=3.8.1
sqlparse~=0.5.2
djangorestframework~=3.15.2
django-cors-headers~=4.6.0
django-filter~=24.3